SDF draw engine FXAA, Cycles SDF fixes, swap HIP kernels for CUDA/OptiX
- Add FXAA fragment shader to SDF draw engine, fix Cycles SDF kernel

+++ intern/cycles/scene/geometry.cpp
@@ -1151,7 +1151,7 @@ void GeometryManager::device_update(Device *device,
         dscene->sdf_shape_atlas.alloc(first_sdf->shape_atlas_data.size());
         memcpy(dscene->sdf_shape_atlas.data(),
                first_sdf->shape_atlas_data.data(),
-               first_sdf->shape_atlas_data.size() * sizeof(float));
+               first_sdf->shape_atlas_data.size() * sizeof(uint16_t));
         dscene->sdf_shape_atlas.copy_to_device();
       }
